% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cn_reset_request_config}
\alias{cn_reset_request_config}
\title{Reset request configuration}
\usage{
cn_reset_request_config()
}
\value{
Invisible request config list.
}
\description{
Reset request configuration
}
