% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cn_set_source}
\alias{cn_set_source}
\title{Set default data source}
\usage{
cn_set_source(
  source = c("auto", "eastmoney", "sina", "netease"),
  fallback_sources = NULL
)
}
\arguments{
\item{source}{One of auto, eastmoney, sina, netease.}

\item{fallback_sources}{Optional source vector used when source is auto.}
}
\value{
Invisible source config list.
}
\description{
Set default data source
}
