% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstockR.R
\name{cn_to_parquet}
\alias{cn_to_parquet}
\title{将数据保存为 parquet 文件}
\usage{
cn_to_parquet(x, file)
}
\arguments{
\item{x}{数据框或 tibble。}

\item{file}{输出 parquet 文件路径。}
}
\value{
隐式返回输出文件路径。
}
\description{
将数据保存为 parquet 文件
}
