% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cn_set_request_config}
\alias{cn_set_request_config}
\title{Update request configuration}
\usage{
cn_set_request_config(
  user_agents = NULL,
  referer = NULL,
  cookie = NULL,
  proxy_pool = NULL,
  headers = list(),
  reset = FALSE
)
}
\arguments{
\item{user_agents}{Optional character vector of User-Agent candidates.}

\item{referer}{Optional single string Referer header.}

\item{cookie}{Optional single string Cookie header.}

\item{proxy_pool}{Optional character vector of proxy URLs.}

\item{headers}{Optional named list of extra headers.}

\item{reset}{Logical flag. If TRUE, reset to defaults.}
}
\value{
Invisible request config list.
}
\description{
Update request configuration
}
